requirements.txt
streamlit
python-dotenv
pymongo
sentence_transformers
google-generativeai==0.6.0
IPython
pydantic==2.7.4
openai==1.35.3